my_bt_lab_institutional_starter_new/my_bt_lab_institutional_starter/requirements.txt
backtrader>=1.9.78.123
pandas>=2.0
PyYAML>=6.0
tushare>=1.4.0
streamlit>=1.31.0
PySide6>=6.7
openpyxl>=3.1
psycopg2-binary>=2.9
sshtunnel>=0.4,<0.5
paramiko<4
